+++ valhalla/userrequests/test/test_api.py
@@ -116,13 +116,21 @@ class TestUserPostRequestApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         self.user = mixer.blend(User)
         mixer.blend(Profile, user=self.user)
         self.client.force_login(self.user)
-        semester = mixer.blend(Semester, id='2016B', start=datetime(2016, 9, 1, tzinfo=timezone.utc),
-                               end=datetime(2016, 12, 31, tzinfo=timezone.utc)
-                               )
-        self.time_allocation_1m0 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=semester,
-                                               telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
-                                               too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
-                                               ipp_time_available=5.0)
+        self.semester = mixer.blend(
+            Semester, id='2016B', start=datetime(2016, 9, 1, tzinfo=timezone.utc),
+            end=datetime(2016, 12, 31, tzinfo=timezone.utc)
+        )
+        self.time_allocation_1m0_sbig = mixer.blend(
+            TimeAllocation, proposal=self.proposal, semester=self.semester,
+            telescope_class='1m0', instrument_name='1M0-SCICAM-SBIG', std_allocation=100.0, std_time_used=0.0,
+            too_allocation=10, too_time_used=0.0, ipp_limit=10.0, ipp_time_available=5.0
+        )
+
+        self.time_allocation_2m0_floyds = mixer.blend(
+            TimeAllocation, proposal=self.proposal, semester=self.semester,
+            telescope_class='2m0', instrument_name='2M0-FLOYDS-SCICAM', std_allocation=100.0, std_time_used=0.0,
+            too_allocation=10, too_time_used=0.0, ipp_limit=10.0, ipp_time_available=5.0
+        )
 
         self.membership = mixer.blend(Membership, user=self.user, proposal=self.proposal)
         self.generic_payload = copy.deepcopy(generic_payload)
@@ -172,17 +180,20 @@ class TestUserPostRequestApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         self.assertEqual(response.status_code, 400)
 
     def test_post_userrequest_no_time_allocation_for_instrument(self):
+        self.time_allocation_2m0_floyds.delete()
         bad_data = self.generic_payload.copy()
         bad_data['requests'][0]['location']['telescope_class'] = '2m0'
-        bad_data['requests'][0]['molecules'][0]['telescope_name'] = '2M0-FLOYDS-SCICAM'
+        bad_data['requests'][0]['molecules'][0]['instrument_name'] = '2M0-FLOYDS-SCICAM'
+        bad_data['requests'][0]['molecules'][0]['type'] = 'SPECTRUM'
+        bad_data['requests'][0]['molecules'][0]['spectra_slit'] = 'slit_6.0as'
         response = self.client.post(reverse('api:user_requests-list'), data=bad_data)
         self.assertEqual(response.status_code, 400)
         self.assertIn('You do not have sufficient time', str(response.content))
 
     def test_post_userrequest_not_enough_time_allocation_for_instrument(self):
         bad_data = self.generic_payload.copy()
-        self.time_allocation_1m0.std_time_used = 99.99
-        self.time_allocation_1m0.save()
+        self.time_allocation_1m0_sbig.std_time_used = 99.99
+        self.time_allocation_1m0_sbig.save()
         response = self.client.post(reverse('api:user_requests-list'), data=bad_data)
         self.assertEqual(response.status_code, 400)
         self.assertIn('does not have enough time allocated', str(response.content))
@@ -190,16 +201,16 @@ class TestUserPostRequestApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
     def test_post_userrequest_not_enough_too_time_allocation_for_instrument(self):
         bad_data = self.generic_payload.copy()
         bad_data['observation_type'] = UserRequest.TOO
-        self.time_allocation_1m0.too_time_used = 9.99
-        self.time_allocation_1m0.save()
+        self.time_allocation_1m0_sbig.too_time_used = 9.99
+        self.time_allocation_1m0_sbig.save()
         response = self.client.post(reverse('api:user_requests-list'), data=bad_data)
         self.assertEqual(response.status_code, 400)
         self.assertIn('does not have enough time allocated', str(response.content))
 
     def test_post_userrequest_not_have_any_time_left(self):
         bad_data = self.generic_payload.copy()
-        self.time_allocation_1m0.std_time_used = 120
-        self.time_allocation_1m0.save()
+        self.time_allocation_1m0_sbig.std_time_used = 120
+        self.time_allocation_1m0_sbig.save()
         response = self.client.post(reverse('api:user_requests-list'), data=bad_data)
         self.assertEqual(response.status_code, 400)
         self.assertIn('does not have any time left allocated', str(response.content))
@@ -232,6 +243,7 @@ class TestUserPostRequestApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         self.assertEqual(response.json()['requests'][0]['molecules'][0]['acquire_radius_arcsec'], 0)
 
         # check that default acquire mode is 'wcs' for floyds
+        bad_data['requests'][0]['location']['telescope_class'] = '2m0'
         bad_data['requests'][0]['molecules'][0]['instrument_name'] = '2M0-FLOYDS-SCICAM'
         bad_data['requests'][0]['molecules'][0]['spectra_slit'] = 'slit_6.0as'
         bad_data['requests'][0]['molecules'][0]['type'] = 'SPECTRUM'
@@ -250,6 +262,7 @@ class TestUserPostRequestApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
 
     def test_post_userrequest_acquire_mode_brightest(self):
         bad_data = self.generic_payload.copy()
+        bad_data['requests'][0]['location']['telescope_class'] = '2m0'
         bad_data['requests'][0]['molecules'][0]['instrument_name'] = '2M0-FLOYDS-SCICAM'
         bad_data['requests'][0]['molecules'][0]['type'] = 'SPECTRUM'
         bad_data['requests'][0]['molecules'][0]['spectra_slit'] = 'slit_6.0as'
@@ -335,18 +348,16 @@ class TestUserRequestIPP(ConfigDBTestMixin, SetTimeMixin, APITestCase):
             end=datetime(2016, 12, 31, tzinfo=timezone.utc)
         )
 
-        self.time_allocation_1m0 = mixer.blend(
+        self.time_allocation_1m0_sbig = mixer.blend(
             TimeAllocation, proposal=self.proposal, semester=semester,
-            telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
-            too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
-            ipp_time_available=5.0
+            telescope_class='1m0', instrument_name='1M0-SCICAM-SBIG', std_allocation=100.0, std_time_used=0.0,
+            too_allocation=10, too_time_used=0.0, ipp_limit=10.0, ipp_time_available=5.0
         )
 
-        self.time_allocation_2m0 = mixer.blend(
+        self.time_allocation_2m0_floyds = mixer.blend(
             TimeAllocation, proposal=self.proposal, semester=semester,
-            telescope_class='2m0', std_allocation=100.0, std_time_used=0.0,
-            too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
-            ipp_time_available=5.0
+            telescope_class='2m0', instrument_name='2M0-FLOYDS-SCICAM', std_allocation=100.0, std_time_used=0.0,
+            too_allocation=10, too_time_used=0.0, ipp_limit=10.0, ipp_time_available=5.0
         )
 
         self.generic_payload = copy.deepcopy(generic_payload)
@@ -369,25 +380,25 @@ class TestUserRequestIPP(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         return UserRequest.objects.get(group_id=ur_dict['group_id'])
 
     def test_user_request_debit_ipp_on_creation(self):
-        self.assertEqual(self.time_allocation_1m0.ipp_time_available, 5.0)
+        self.assertEqual(self.time_allocation_1m0_sbig.ipp_time_available, 5.0)
 
         ur = self.generic_payload.copy()
         response = self.client.post(reverse('api:user_requests-list'), data=ur)
         self.assertEqual(response.status_code, 201)
 
         # verify that now that the object is saved, ipp has been debited
-        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0.id)
+        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0_sbig.id)
         self.assertLess(time_allocation.ipp_time_available, 5.0)
 
     def test_user_request_credit_ipp_on_cancelation(self):
         user_request = self._build_user_request(self.generic_payload.copy())
         # verify that now that the TimeAllocation has been debited
-        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0.id)
+        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0_sbig.id)
         self.assertLess(time_allocation.ipp_time_available, 5.0)
         user_request.state = 'CANCELED'
         user_request.save()
         # verify that now that the TimeAllocation has its original ipp value
-        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0.id)
+        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0_sbig.id)
         self.assertEqual(time_allocation.ipp_time_available, 5.0)
         # also verify that the child request state has changed to window_expired as well
         self.assertEqual(user_request.requests.first().state, 'CANCELED')
@@ -395,18 +406,18 @@ class TestUserRequestIPP(ConfigDBTestMixin, SetTimeMixin, APITestCase):
     def test_user_request_credit_ipp_on_expiration(self):
         user_request = self._build_user_request(self.generic_payload.copy())
         # verify that now that the TimeAllocation has been debited
-        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0.id)
+        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0_sbig.id)
         self.assertLess(time_allocation.ipp_time_available, 5.0)
         user_request.state = 'WINDOW_EXPIRED'
         user_request.save()
         # verify that now that the TimeAllocation has its original ipp value
-        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0.id)
+        time_allocation = TimeAllocation.objects.get(pk=self.time_allocation_1m0_sbig.id)
         self.assertEqual(time_allocation.ipp_time_available, 5.0)
         # also verify that the child request state has changed to window_expired as well
         self.assertEqual(user_request.requests.first().state, 'WINDOW_EXPIRED')
 
     def test_user_request_debit_ipp_on_creation_fail(self):
-        self.assertEqual(self.time_allocation_1m0.ipp_time_available, 5.0)
+        self.assertEqual(self.time_allocation_1m0_sbig.ipp_time_available, 5.0)
 
         ur = self.generic_payload.copy()
         # ipp value that is too high, will be rejected
@@ -414,6 +425,7 @@ class TestUserRequestIPP(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         response = self.client.post(reverse('api:user_requests-list'), data=ur)
         self.assertEqual(response.status_code, 400)
         self.assertIn('TimeAllocationError', str(response.content))
+        self.assertIn('ipp_value of 100.0 requires more ipp_time then is available.', str(response.content))
 
         # verify that objects were not created by the send
         self.assertFalse(UserRequest.objects.filter(group_id='ipp_request').exists())
@@ -423,9 +435,9 @@ class TestUserRequestIPP(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         ur['operator'] = 'MANY'
         user_request = self._build_user_request(ur)
         # verify that now that both the TimeAllocation has been debited
-        time_allocation_1m0 = TimeAllocation.objects.get(pk=self.time_allocation_1m0.id)
+        time_allocation_1m0 = TimeAllocation.objects.get(pk=self.time_allocation_1m0_sbig.id)
         self.assertLess(time_allocation_1m0.ipp_time_available, 5.0)
-        time_allocation_2m0 = TimeAllocation.objects.get(pk=self.time_allocation_2m0.id)
+        time_allocation_2m0 = TimeAllocation.objects.get(pk=self.time_allocation_2m0_floyds.id)
         self.assertLess(time_allocation_2m0.ipp_time_available, 5.0)
         # now set one request to completed, then set the user request to unschedulable
         request = user_request.requests.first()
@@ -434,9 +446,9 @@ class TestUserRequestIPP(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         user_request.state = 'WINDOW_EXPIRED'
         user_request.save()
         # now verify that time allocation 1 is still debited, but time allocation 2 has been credited back its time
-        time_allocation_1m0 = TimeAllocation.objects.get(pk=self.time_allocation_1m0.id)
+        time_allocation_1m0 = TimeAllocation.objects.get(pk=self.time_allocation_1m0_sbig.id)
         self.assertLess(time_allocation_1m0.ipp_time_available, 5.0)
-        time_allocation_2m0 = TimeAllocation.objects.get(pk=self.time_allocation_2m0.id)
+        time_allocation_2m0 = TimeAllocation.objects.get(pk=self.time_allocation_2m0_floyds.id)
         self.assertEqual(time_allocation_2m0.ipp_time_available, 5.0)
 
 
@@ -458,9 +470,8 @@ class TestRequestIPP(ConfigDBTestMixin, SetTimeMixin, APITestCase):
 
         self.time_allocation_1m0 = mixer.blend(
             TimeAllocation, proposal=self.proposal, semester=semester,
-            telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
-            too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
-            ipp_time_available=5.0
+            telescope_class='1m0', instrument_name='1M0-SCICAM-SBIG', std_allocation=100.0, std_time_used=0.0,
+            too_allocation=10, too_time_used=0.0, ipp_limit=10.0, ipp_time_available=5.0
         )
 
         self.generic_payload = copy.deepcopy(generic_payload)
@@ -617,7 +628,7 @@ class TestCadenceApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
             TimeAllocation, proposal=self.proposal, semester=semester,
             telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
             too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
-            ipp_time_available=5.0
+            ipp_time_available=5.0, instrument_name='1M0-SCICAM-SBIG'
         )
 
         self.client.force_login(self.user)
@@ -707,11 +718,18 @@ class TestSiderealTarget(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         semester = mixer.blend(Semester, id='2016B', start=datetime(2016, 9, 1, tzinfo=timezone.utc),
                                end=datetime(2016, 12, 31, tzinfo=timezone.utc)
                                )
-        self.time_allocation_1m0 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=semester,
-                                               telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
-                                               too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
-                                               ipp_time_available=5.0
-                                               )
+        self.time_allocation_1m0_sbig = mixer.blend(
+            TimeAllocation, proposal=self.proposal, semester=semester,
+            telescope_class='1m0', instrument_name='1M0-SCICAM-SBIG', std_allocation=100.0,
+            std_time_used=0.0, too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
+            ipp_time_available=5.0
+        )
+
+        self.time_allocation_2m0_floyds = mixer.blend(
+            TimeAllocation, proposal=self.proposal, semester=semester,
+            telescope_class='2m0', instrument_name='2M0-FLOYDS-SCICAM', std_allocation=100.0, std_time_used=0.0,
+            too_allocation=10, too_time_used=0.0, ipp_limit=10.0, ipp_time_available=5.0
+        )
 
         mixer.blend(Membership, user=self.user, proposal=self.proposal)
         self.generic_payload = copy.deepcopy(generic_payload)
@@ -764,6 +782,7 @@ class TestSiderealTarget(ConfigDBTestMixin, SetTimeMixin, APITestCase):
 
     def test_floyds_gets_vfloat_default(self):
         good_data = self.generic_payload.copy()
+        good_data['requests'][0]['location']['telescope_class'] = '2m0'
         good_data['requests'][0]['molecules'][0]['instrument_name'] = '2M0-FLOYDS-SCICAM'
         good_data['requests'][0]['molecules'][0]['type'] = 'SPECTRUM'
         good_data['requests'][0]['molecules'][0]['spectra_slit'] = 'slit_6.0as'
@@ -789,6 +808,7 @@ class TestNonSiderealTarget(ConfigDBTestMixin, SetTimeMixin, APITestCase):
                                end=datetime(2016, 12, 31, tzinfo=timezone.utc))
         self.time_allocation_1m0 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=semester,
                                                telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
+                                               instrument_name='1M0-SCICAM-SBIG',
                                                too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
                                                ipp_time_available=5.0)
 
@@ -895,6 +915,7 @@ class TestSatelliteTarget(ConfigDBTestMixin, SetTimeMixin, APITestCase):
 
         self.time_allocation_1m0 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=semester,
                                                telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
+                                               instrument_name='1M0-SCICAM-SBIG',
                                                too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
                                                ipp_time_available=5.0)
 
@@ -944,7 +965,11 @@ class TestLocationApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         self.time_allocation_1m0 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=semester,
                                                telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
                                                too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
-                                               ipp_time_available=5.0)
+                                               ipp_time_available=5.0, instrument_name='1M0-SCICAM-SBIG')
+        self.time_allocation_2m0 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=semester,
+                                               telescope_class='2m0', std_allocation=100.0, std_time_used=0.0,
+                                               too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
+                                               ipp_time_available=5.0, instrument_name='2M0-FLOYDS-SCICAM')
 
         mixer.blend(Membership, user=self.user, proposal=self.proposal)
         self.generic_payload = copy.deepcopy(generic_payload)
@@ -1003,6 +1028,13 @@ class TestLocationApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         self.assertEqual(response.status_code, 201)
         self.assertIsNone(response.json()['requests'][0]['location'].get('observatory'))
 
+    def test_post_userrequest_location_instrument_doesnt_match_class(self):
+        good_data = self.generic_payload.copy()
+        good_data['requests'][0]['location']['telescope_class'] = '2m0'
+        good_data['requests'][0]['molecules'][0]['instrument_name'] = '1M0-SCICAM-SBIG'
+        response = self.client.post(reverse('api:user_requests-list'), data=good_data)
+        self.assertNotEqual(response.status_code, 201)
+
 
 class TestMoleculeApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
     def setUp(self):
@@ -1013,10 +1045,20 @@ class TestMoleculeApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
 
         semester = mixer.blend(Semester, id='2016B', start=datetime(2016, 9, 1, tzinfo=timezone.utc),
                                end=datetime(2016, 12, 31, tzinfo=timezone.utc))
-        self.time_allocation_1m0 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=semester,
-                                               telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
-                                               too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
-                                               ipp_time_available=5.0)
+
+        self.time_allocation_1m0_sbig = mixer.blend(
+            TimeAllocation, proposal=self.proposal, semester=semester,
+            telescope_class='1m0', instrument_name='1M0-SCICAM-SBIG', std_allocation=100.0,
+            std_time_used=0.0, too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
+            ipp_time_available=5.0
+        )
+
+        self.time_allocation_2m0_floyds = mixer.blend(
+            TimeAllocation, proposal=self.proposal, semester=semester,
+            telescope_class='2m0', instrument_name='2M0-FLOYDS-SCICAM', std_allocation=100.0,
+            std_time_used=0.0, too_allocation=10, too_time_used=0.0, ipp_limit=10.0,
+            ipp_time_available=5.0
+        )
 
         mixer.blend(Membership, user=self.user, proposal=self.proposal)
         self.generic_payload = copy.deepcopy(generic_payload)
@@ -1032,6 +1074,7 @@ class TestMoleculeApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         self.assertEqual(molecule['ag_mode'], 'OPTIONAL')
         self.assertEqual(molecule['spectra_slit'], '')
 
+        good_data['requests'][0]['location']['telescope_class'] = '2m0'
         good_data['requests'][0]['molecules'][0]['instrument_name'] = '2M0-FLOYDS-SCICAM'
         good_data['requests'][0]['molecules'][0]['spectra_slit'] = 'slit_6.0as'
         good_data['requests'][0]['molecules'][0]['type'] = 'LAMP_FLAT'
@@ -1051,6 +1094,8 @@ class TestMoleculeApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
 
     def test_filter_not_necessary_for_type(self):
         good_data = self.generic_payload.copy()
+
+        good_data['requests'][0]['location']['telescope_class'] = '2m0'
         good_data['requests'][0]['molecules'][0]['type'] = 'ARC'
         good_data['requests'][0]['molecules'][0]['instrument_name'] = '2M0-FLOYDS-SCICAM'
         good_data['requests'][0]['molecules'][0]['spectra_slit'] = 'slit_6.0as'
@@ -2008,14 +2053,16 @@ class TestMaxIppUserrequestApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         self.semester = mixer.blend(Semester, id='2016B', start=datetime(2016, 9, 1, tzinfo=timezone.utc),
                                     end=datetime(2016, 12, 31, tzinfo=timezone.utc))
 
-        self.time_allocation_1m0 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=self.semester,
-                                               telescope_class='1m0', std_allocation=100.0, std_time_used=0.0,
-                                               too_allocation=10.0, too_time_used=0.0, ipp_limit=10.0,
-                                               ipp_time_available=1.0)
-        self.time_allocation_0m4 = mixer.blend(TimeAllocation, proposal=self.proposal, semester=self.semester,
-                                               telescope_class='0m4', std_allocation=100.0, std_time_used=0.0,
-                                               too_allocation=10.0, too_time_used=0.0, ipp_limit=10.0,
-                                               ipp_time_available=1.0)
+        self.time_allocation_1m0_sbig = mixer.blend(
+            TimeAllocation, proposal=self.proposal, semester=self.semester, telescope_class='1m0',
+            instrument_name='1M0-SCICAM-SBIG', std_allocation=100.0, std_time_used=0.0,
+            too_allocation=10.0, too_time_used=0.0, ipp_limit=10.0, ipp_time_available=1.0
+        )
+        self.time_allocation_0m4_sbig = mixer.blend(
+            TimeAllocation, proposal=self.proposal, semester=self.semester, telescope_class='0m4',
+            instrument_name='0M4-SCICAM-SBIG', std_allocation=100.0, std_time_used=0.0,
+            too_allocation=10.0, too_time_used=0.0, ipp_limit=10.0, ipp_time_available=1.0
+        )
         self.user = mixer.blend(User)
         mixer.blend(Membership, user=self.user, proposal=self.proposal)
         self.client.force_login(self.user)
@@ -2036,8 +2083,12 @@ class TestMaxIppUserrequestApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
 
         ipp_dict = response.json()
         self.assertIn(self.semester.id, ipp_dict)
-        self.assertEqual(MAX_IPP_LIMIT, ipp_dict[self.semester.id]['1m0']['max_allowable_ipp_value'])
-        self.assertEqual(MIN_IPP_LIMIT, ipp_dict[self.semester.id]['1m0']['min_allowable_ipp_value'])
+        self.assertEqual(
+            MAX_IPP_LIMIT, ipp_dict[self.semester.id]['1m0']['1M0-SCICAM-SBIG']['max_allowable_ipp_value']
+        )
+        self.assertEqual(
+            MIN_IPP_LIMIT, ipp_dict[self.semester.id]['1m0']['1M0-SCICAM-SBIG']['min_allowable_ipp_value']
+        )
 
     def test_get_max_ipp_reduced_max_ipp(self):
 
@@ -2048,19 +2099,18 @@ class TestMaxIppUserrequestApi(ConfigDBTestMixin, SetTimeMixin, APITestCase):
         ipp_dict = response.json()
         self.assertIn(self.semester.id, ipp_dict)
         # max ipp allowable is close to 1.0 ipp_available / 1.5 ~duration + 1.
-        self.assertEqual(1.649, ipp_dict[self.semester.id]['1m0']['max_allowable_ipp_value'])
+        self.assertEqual(1.649, ipp_dict[self.semester.id]['1m0']['1M0-SCICAM-SBIG']['max_allowable_ipp_value'])
 
     def test_get_max_ipp_no_ipp_available(self):
         good_data = self.generic_payload.copy()
-        self.time_allocation_1m0.ipp_time_available = 0.0
-        self.time_allocation_1m0.save()
+        self.time_allocation_1m0_sbig.ipp_time_available = 0.0
+        self.time_allocation_1m0_sbig.save()
         response = self.client.post(reverse('api:user_requests-max-allowable-ipp'), good_data)
         self.assertEqual(response.status_code, 200)
         ipp_dict = response.json()
         self.assertIn(self.semester.id, ipp_dict)
         # max ipp allowable is close to 1.0 ipp_available / 1.5 ~duration + 1.
-        self.assertEqual(1.0, ipp_dict[self.semester.id]['1m0']['max_allowable_ipp_value'])
-
+        self.assertEqual(1.0, ipp_dict[self.semester.id]['1m0']['1M0-SCICAM-SBIG']['max_allowable_ipp_value'])
 
 class TestFiltering(APITestCase):
     def test_filtering_works(self):